% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbi_api_agencies}
\alias{fbi_api_agencies}
\title{All agencies included in the FBI's Crime Data Explorer API.}
\format{
A data frame with 18459 rows and 13 variables:
\describe{
  \item{ori}{9-digit unique ID number for the agency}
  \item{agency_name}{Agency name}
  \item{agency_type_name}{Type of department - e.g city, county, school}
  \item{state_name}{State name}
  \item{state_abbr}{State abbreviation}
  \item{division_name}{Census division name}
  \item{region_name}{Census region name}
  \item{region_desc}{Census region code}
  \item{county_name}{County name}
  \item{nibrs}{Boolean on whether the agency reports to NBIRS}
  \item{latitude}{Agency latitude}
  \item{longitude}{Agency longitude}
  \item{nibrs_start_date}{Date agency started reporting to NIBRS. NULL if
  the agency does not report to NIBRS}

}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/}
}
\usage{
fbi_api_agencies
}
\description{
A dataset containing information about the 18,459 agencies included in the
FBI's API.
}
\keyword{datasets}
