% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{flatten_cde_json}
\alias{flatten_cde_json}
\title{Flatten CDE JSON response}
\usage{
flatten_cde_json(obj)
}
\arguments{
\item{obj}{A nested list/object from the CDE API with structure
\code{label -> period -> value}}
}
\value{
A data.frame with columns: label, period, value
}
\description{
Turns an object-of-\code{label -> period -> value} into a long/tidy data.frame
}
\examples{
obj <- list(
  "United States Offenses" = list(
    "01-2015" = 27.6,
    "01-2016" = 29.62
  )
)
\dontrun{
flatten_cde_json(obj)
}
}
