% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_arrest.R
\name{get_arrest_demographics}
\alias{get_arrest_demographics}
\title{Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set.}
\usage{
get_arrest_demographics(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  offense = "all",
  end_year = make_year(),
  key = get_api_key()
)
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}

\item{offense}{A string or vector of strings with the offenses you want to scrape. If
input is 'all' (default), returns data for all offenses. Please run `list_ucr_arrest_offenses()` to see all possible offenses.}
}
\value{
A data.frame with the number of arrests by arrestee demographic for each crime-year in the jurisdiction.
}
\description{
Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set.
}
\examples{
\dontrun{
get_arrest_demographics(ori = "CA0010900", offense = "robbery")
get_arrest_demographics(offense = "robbery")
}
}
