% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_arrest.R
\name{get_arrest_demographics_all}
\alias{get_arrest_demographics_all}
\title{Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set for all offenses.}
\usage{
get_arrest_demographics_all(...)
}
\arguments{
\item{...}{all parameters valid for \link{get_arrest_demographics}}
}
\value{
A data.frame with the number of arrests by arrestee demographic for each crime-year in the jurisdiction.
}
\description{
Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set for all offenses.
}
\examples{
\dontrun{
get_arrest_demographics_all(region = "South")
}
}
