% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_estimated_arson}
\alias{get_estimated_arson}
\title{Get estimated arson data}
\usage{
get_estimated_arson(state_abb = NULL, from = "01-2015", to = "12-2020")
}
\arguments{
\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}

\item{from}{Start date in MM-YYYY format (default "01-2015").}

\item{to}{End date in MM-YYYY format (default "12-2020").}
}
\value{
A data.frame with columns: geography, offense, period, count, rate
}
\description{
Get estimated arson data
}
\examples{
\dontrun{
get_estimated_arson("CA")
}
}
\seealso{
Other UCR crime functions:
\code{\link[=get_agency_crime]{get_agency_crime()}},
\code{\link[=get_agency_participation]{get_agency_participation()}},
\code{\link[=get_estimated_crime]{get_estimated_crime()}},
\code{\link[=get_region_participation]{get_region_participation()}},
\code{\link[=get_state_participation]{get_state_participation()}}
}
\concept{UCR crime functions}
