% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_participation_region}
\alias{get_participation_region}
\title{Get regional level UCR and NIBRS participation data}
\usage{
get_participation_region(region_name = NULL, key = get_api_key())
}
\arguments{
\item{key}{A string containing your FBI's Crime Data Explorer API key}

\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data.}
}
\value{
A data.frame with regional NIBRS and UCR participation rates for selected state#'
}
\description{
Get regional level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_participation_region("South")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_agency_crime}()},
\code{\link{get_estimated_arson}()},
\code{\link{get_estimated_crime}()},
\code{\link{get_participation_state}()}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
