% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_participation_state}
\alias{get_participation_state}
\title{Get state level UCR and NIBRS participation data}
\usage{
get_participation_state(state_abb = NULL, key = get_api_key())
}
\arguments{
\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with state-level NIBRS and UCR participation rates for selected state
}
\description{
Get state level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_participation_state("MT")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_agency_crime}()},
\code{\link{get_estimated_arson}()},
\code{\link{get_estimated_crime}()},
\code{\link{get_participation_region}()}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
