% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{is_valid_state}
\alias{is_valid_state}
\title{Validate a state abbreviation}
\usage{
is_valid_state(state_abb)
}
\arguments{
\item{state_abb}{Character string or vector of state abbreviations.}
}
\value{
Logical vector, TRUE for valid state abbreviations.
}
\description{
Checks if a state abbreviation is a valid US state or territory code.
}
\examples{
is_valid_state("CA")
is_valid_state("XX")
}
