% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_regions}
\alias{list_regions}
\title{Return a vector of all regions choices.}
\usage{
list_regions()
}
\value{
A vector of all regions choices.
}
\description{
Return a vector of all regions choices.
}
\examples{
list_regions()
}
\seealso{
Other NIBRS functions:
\code{\link[=get_nibrs_offender]{get_nibrs_offender()}},
\code{\link[=get_nibrs_offense]{get_nibrs_offense()}},
\code{\link[=get_nibrs_victim]{get_nibrs_victim()}},
\code{\link[=list_nibrs_offender_variables]{list_nibrs_offender_variables()}},
\code{\link[=list_nibrs_offense_variables]{list_nibrs_offense_variables()}},
\code{\link[=list_nibrs_offenses]{list_nibrs_offenses()}},
\code{\link[=list_nibrs_victim_variables]{list_nibrs_victim_variables()}},
\code{\link[=list_ucr_arrest_offenses]{list_ucr_arrest_offenses()}}
}
\concept{NIBRS functions}
