% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{cde_base_url}
\alias{cde_base_url}
\title{CDE base URL}
\usage{
cde_base_url()
}
\value{
A character string with the base URL.
}
\description{
Returns the base URL for the FBI CDE API. Overridable via the
`fbi_cde_base_url` option or the `FBI_CDE_BASE_URL` environment variable.
}
\examples{
cde_base_url()
}
