% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_agency_crime}
\alias{get_agency_crime}
\title{Get agency-level crime data from the UCR Offenses Known and Clearances}
\usage{
get_agency_crime(ori, from = "01-2015", to = "12-2020", offense = "V")
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{from}{Start date in MM-YYYY format (default "01-2015").}

\item{to}{End date in MM-YYYY format (default "12-2020").}

\item{offense}{Offense code (default "V" for violent crime). See
`get_offense_codes()` for available codes.}
}
\value{
A data.frame with columns: geography, offense, period, count, rate
}
\description{
Get agency-level crime data from the UCR Offenses Known and Clearances
}
\examples{
\dontrun{
get_agency_crime("AK0010100")
}
}
\seealso{
Other UCR crime functions:
\code{\link[=get_agency_participation]{get_agency_participation()}},
\code{\link[=get_estimated_arson]{get_estimated_arson()}},
\code{\link[=get_estimated_crime]{get_estimated_crime()}},
\code{\link[=get_leoka]{get_leoka()}},
\code{\link[=get_leoka_monthly]{get_leoka_monthly()}},
\code{\link[=get_region_participation]{get_region_participation()}},
\code{\link[=get_state_participation]{get_state_participation()}}
}
\concept{UCR crime functions}
