% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_agency_crime}
\alias{get_agency_crime}
\title{Get agency-level crime data from the UCR's Offenses Known and Clearances by Arrest data set.}
\usage{
get_agency_crime(ori, key = get_api_key())
}
\arguments{
\item{ori}{A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with agency-level UCR crime data for selected agency.
}
\description{
Get agency-level crime data from the UCR's Offenses Known and Clearances by Arrest data set.
}
\examples{
\dontrun{
get_agency_crime("AK0010100")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_estimated_arson}()},
\code{\link{get_estimated_crime}()},
\code{\link{get_participation_region}()},
\code{\link{get_participation_state}()}
}
\concept{UCR crime functions}
