% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_agency_participation}
\alias{get_agency_participation}
\title{Get agency level UCR and NIBRS participation data}
\usage{
get_agency_participation(ori, key = get_api_key())
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}
}
\value{
A data.frame with regional NIBRS and UCR participation rates for selected agency
}
\description{
Get agency level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_agency_participation("AK0010100")
}
}
\seealso{
Other UCR crime functions:
\code{\link[=get_agency_crime]{get_agency_crime()}},
\code{\link[=get_estimated_arson]{get_estimated_arson()}},
\code{\link[=get_estimated_crime]{get_estimated_crime()}},
\code{\link[=get_region_participation]{get_region_participation()}},
\code{\link[=get_state_participation]{get_state_participation()}}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
