% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_agency_participation}
\alias{get_agency_participation}
\title{Get agency level UCR and NIBRS participation data}
\usage{
get_agency_participation(ori, key = get_api_key())
}
\arguments{
\item{ori}{A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with regional NIBRS and UCR participation rates for selected agency
}
\description{
Get agency level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_agency_participation("AK0010100")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_agency_crime}()},
\code{\link{get_estimated_arson}()},
\code{\link{get_estimated_crime}()},
\code{\link{get_region_participation}()},
\code{\link{get_state_participation}()}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
