% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_arrest.R
\name{get_arrest_count}
\alias{get_arrest_count}
\title{Get arrest data from the UCR's Arrests by Age, Sex, and Race data set.}
\usage{
get_arrest_count(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  monthly = FALSE,
  end_year = make_year(),
  key = get_api_key()
)
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}

\item{end_year}{= last year of the collection to get, defaults to the prior calendar year from today
If TRUE (not default), returns data as monthly units. Otherwise returns annual data.}
}
\value{
A data.frame with the number of arrests for each crime-year in the jurisdiction.
}
\description{
Get arrest data from the UCR's Arrests by Age, Sex, and Race data set.
}
\examples{
\dontrun{
get_arrest_count(ori = "CA0010900")
get_arrest_count()
}
}
