% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_arrest.R
\name{get_arrest_demographics}
\alias{get_arrest_demographics}
\title{Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set.}
\usage{
get_arrest_demographics(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  offense = "all",
  end_year = make_year(),
  key = get_api_key()
)
}
\arguments{
\item{ori}{A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.}

\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.}

\item{region}{String or vector of strings input for region name(s) to get data for. Please run `list_regions()`  to see all possible variables.}

\item{offense}{A string or vector of strings with the offenses you want to scrape. If
input is 'all' (default), returns data for all offenses. Please run `list_ucr_arrest_offenses()` to see all possible offenses.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with the number of arrests by arrestee demographic for each crime-year in the jurisdiction.
}
\description{
Get arrest data by arrestee demographic from the UCR's Arrests by Age, Sex, and Race data set.
}
\examples{
\dontrun{
get_arrest_demographics(ori = "CA0010900", offense = "robbery")
get_arrest_demographics(offense = "robbery")
}
}
