% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_crime.R
\name{get_estimated_arson}
\alias{get_estimated_arson}
\title{Get state-, region- or national-level estimated arson data from the UCR's Offenses Known and Clearances by Arrest data set.}
\usage{
get_estimated_arson(state_abb = NULL, region = NULL, key = get_api_key())
}
\arguments{
\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.}

\item{region}{String or vector of strings input for region name(s) to get data for. Please run `list_regions()`  to see all possible variables.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with state-level estimated UCR arson data for selected state
}
\description{
Get state-, region- or national-level estimated arson data from the UCR's Offenses Known and Clearances by Arrest data set.
}
\examples{
\dontrun{
get_estimated_arson("CA")
}
}
\seealso{
Other UCR crime functions: 
\code{\link{get_agency_crime}()},
\code{\link{get_estimated_crime}()},
\code{\link{get_participation_region}()},
\code{\link{get_participation_state}()}
}
\concept{UCR crime functions}
