% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{get_nibrs_offender}
\alias{get_nibrs_offender}
\title{Gets offender-level data from the FBI's National Incident-Based Reporting System (NIBRS)}
\usage{
get_nibrs_offender(
  key = get_api_key(),
  offense = "all",
  variable = "count",
  ori = NULL,
  region = NULL,
  state_abb = NULL
)
}
\arguments{
\item{offense}{A string or vector of strings with the offenses you want to scrape. If
input is 'all' (default), returns data for all offenses. Please run `list_nibrs_offenses()` to see all possible offenses.}

\item{variable}{A string with the variable you want the data to be in. This is essentially the units you want returned (number of offenders by age, race, sex, etc.). By default returns total count of offenders Please run `list_nibrs_offender_variables()`  to see all possible variables.}

\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}
}
\value{
A data.frame of NIBRS offender data.
}
\description{
Gets offender-level data from the FBI's National Incident-Based Reporting System (NIBRS)
}
\examples{
\dontrun{
get_nibrs_offender(ori = "DE0010100")
}
}
\seealso{
Other NIBRS functions:
\code{\link[=get_nibrs_offense]{get_nibrs_offense()}},
\code{\link[=get_nibrs_victim]{get_nibrs_victim()}},
\code{\link[=list_nibrs_offender_variables]{list_nibrs_offender_variables()}},
\code{\link[=list_nibrs_offense_variables]{list_nibrs_offense_variables()}},
\code{\link[=list_nibrs_offenses]{list_nibrs_offenses()}},
\code{\link[=list_nibrs_victim_variables]{list_nibrs_victim_variables()}},
\code{\link[=list_regions]{list_regions()}}
}
\concept{NIBRS functions}
