% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_police_employment.R
\name{get_police_employment}
\alias{get_police_employment}
\title{Get agency-, state-, region-, or national-level police staffing data.}
\usage{
get_police_employment(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  start_year = 1985,
  end_year = make_year() - 2,
  key = get_api_key()
)
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}
}
\value{
A data.frame with columns for annual number of employees and officers (also broken up by gender).
}
\description{
Get agency-, state-, region-, or national-level police staffing data.
}
\examples{
\dontrun{
# Gets only Oakland Police Department in California
get_police_employment("CA0010900")

# Gets California state-level estimates
get_police_employment(state_abb = "CA")

# Gets national-level estimates
get_police_employment()
}
}
