% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_employment.R
\name{get_police_employment}
\alias{get_police_employment}
\title{Get agency-, state-, region-, or national-level police staffing data.}
\usage{
get_police_employment(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  from = "2015",
  to = "2020"
)
}
\arguments{
\item{ori}{A string of the 9-character ORI code for the desired agency.}

\item{state_abb}{String for state abbreviation. If `NULL` (default) returns
national data.}

\item{region}{Character string for the census region (e.g. "Northeast",
"Midwest", "South", "West").}

\item{from}{Start date in MM-YYYY format (default "01-2015").}

\item{to}{End date in MM-YYYY format (default "12-2020").}
}
\value{
A data.frame with columns for annual number of employees and officers
  (also broken up by gender).
}
\description{
Get agency-, state-, region-, or national-level police staffing data.
}
\examples{
\dontrun{
# Gets only Oakland Police Department in California
get_police_employment("CA0010900")

# Gets California state-level estimates
get_police_employment(state_abb = "CA")

# Gets national-level estimates
get_police_employment()
}
}
