% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_police_employment.R
\name{get_police_employment}
\alias{get_police_employment}
\title{Get agency-, state-, region-, or national-level police staffing data.}
\usage{
get_police_employment(
  ori = NULL,
  state_abb = NULL,
  region = NULL,
  end_year = make_year() - 2,
  key = get_api_key()
)
}
\arguments{
\item{ori}{A string or vector of strings of the 9-character ORI code (unique agency ID) for the desired agency.}

\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data. If `state_abb` and `region` both have values, will use return `region` input data.}

\item{region}{String or vector of strings input for region name(s) to get data for. Please run `list_regions()`  to see all possible variables.}

\item{key}{A string containing your FBI's Crime Data Explorer API key}
}
\value{
A data.frame with columns for annual number of employees and officers (also broken up by gender).
}
\description{
Get agency-, state-, region-, or national-level police staffing data.
}
\examples{
\dontrun{
# Gets only Oakland Police Department in California
get_police_employment("CA0010900")

# Gets California state-level estimates
get_police_employment(state_abb = "CA")

# Gets national-level estimates
get_police_employment()
}
}
