% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_region_participation}
\alias{get_region_participation}
\title{Get regional level UCR and NIBRS participation data}
\usage{
get_region_participation(region_name = NULL, key = get_api_key())
}
\arguments{
\item{region_name}{String or vector of strings for region to get data for}
}
\value{
A data.frame with regional NIBRS and UCR participation rates for selected state#'
}
\description{
Get regional level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_region_participation("South")
}
}
\seealso{
Other UCR crime functions:
\code{\link[=get_agency_crime]{get_agency_crime()}},
\code{\link[=get_agency_participation]{get_agency_participation()}},
\code{\link[=get_estimated_arson]{get_estimated_arson()}},
\code{\link[=get_estimated_crime]{get_estimated_crime()}},
\code{\link[=get_state_participation]{get_state_participation()}}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
