% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_participation.R
\name{get_state_participation}
\alias{get_state_participation}
\title{Get state level UCR and NIBRS participation data}
\usage{
get_state_participation(state_abb = NULL, key = get_api_key())
}
\arguments{
\item{state_abb}{String or vector of strings input for state abbreviation(s) to get data for.
If NULL (default) returns national data.}
}
\value{
A data.frame with state-level NIBRS and UCR participation rates for selected state
}
\description{
Get state level UCR and NIBRS participation data
}
\examples{
\dontrun{
get_state_participation("MT")
}
}
\seealso{
Other UCR crime functions:
\code{\link[=get_agency_crime]{get_agency_crime()}},
\code{\link[=get_agency_participation]{get_agency_participation()}},
\code{\link[=get_estimated_arson]{get_estimated_arson()}},
\code{\link[=get_estimated_crime]{get_estimated_crime()}},
\code{\link[=get_region_participation]{get_region_participation()}}
}
\author{
Jared E. Knowles, Civilytics Consulting
}
\concept{UCR crime functions}
