% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibrs.R
\name{list_nibrs_offenses}
\alias{list_nibrs_offenses}
\title{Return a vector of all offenses in the NIBRS data.}
\usage{
list_nibrs_offenses()
}
\value{
A character vector of NIBRS offense codes.
}
\description{
Return a vector of all offenses in the NIBRS data.
}
\examples{
list_nibrs_offenses()
}
\seealso{
Other NIBRS functions:
\code{\link[=get_nibrs_offender]{get_nibrs_offender()}},
\code{\link[=get_nibrs_offense]{get_nibrs_offense()}},
\code{\link[=get_nibrs_victim]{get_nibrs_victim()}},
\code{\link[=list_nibrs_offender_variables]{list_nibrs_offender_variables()}},
\code{\link[=list_nibrs_offense_variables]{list_nibrs_offense_variables()}},
\code{\link[=list_nibrs_victim_variables]{list_nibrs_victim_variables()}},
\code{\link[=list_regions]{list_regions()}}
}
\concept{NIBRS functions}
