% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_arrest.R
\name{list_ucr_arrest_offenses}
\alias{list_ucr_arrest_offenses}
\title{Return a vector of all offenses in the UCR Arrest data.}
\usage{
list_ucr_arrest_offenses()
}
\value{
A character vector of UCR arrest offense codes.
}
\description{
Return a vector of all offenses in the UCR Arrest data.
}
\examples{
list_ucr_arrest_offenses()
}
\seealso{
Other UCR arrest functions:
\code{\link[=get_arrest_count]{get_arrest_count()}},
\code{\link[=get_arrest_demographics]{get_arrest_demographics()}},
\code{\link[=get_arrest_demographics_all]{get_arrest_demographics_all()}}
}
\concept{UCR arrest functions}
