% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nibrs_offense_variables}
\alias{nibrs_offense_variables}
\title{All available variables for NIBRS offense data.}
\format{
A vector with 8 elements:
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/}
}
\usage{
nibrs_offense_variables
}
\description{
The `get_nibrs_offense()` function requires a string input for
which variable you want data on. This is a vector of strings
with all possible variables.
}
\keyword{datasets}
