% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nibrs_offense_variables}
\alias{nibrs_offense_variables}
\title{All available variables for NIBRS offense data.}
\format{
A character vector with 2 elements.
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/}
}
\usage{
nibrs_offense_variables
}
\description{
The values \code{get_nibrs_offense()} accepts for \code{variable}: the keys of a
NIBRS response's offense section. Built by \code{data-raw/api_vocabularies.R}.
}
\keyword{datasets}
