% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_api_key}
\alias{set_api_key}
\title{Sets the FBI's Crime Data Explorer API key}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{A character string with the API key.}
}
\value{
A string with the FBI's Crime Data Explorer API key
}
\description{
Sets the FBI's Crime Data Explorer API key
}
\examples{
set_api_key("abc123")
}
